[STRIPE]: rename stripeConfig into StripeProperties and move to another package
