+++ src/main/java/com/example/carsharingservice/stripepaymant/StripePaymentProdiverImpl.java
@@ -1,10 +1,8 @@
-package com.example.carsharingservice.service.impl;
+package com.example.carsharingservice.stripepaymant;
 
-import com.example.carsharingservice.config.StripeConfig;
 import com.example.carsharingservice.model.Payment;
 import com.example.carsharingservice.model.PaymentStatus;
 import com.example.carsharingservice.service.PaymentService;
-import com.example.carsharingservice.service.StripePaymentService;
 import com.stripe.Stripe;
 import com.stripe.exception.StripeException;
 import com.stripe.model.Price;
@@ -18,13 +16,13 @@ import org.springframework.stereotype.Service;
 
 @AllArgsConstructor
 @Service
-public class StripePaymentServiceImpl implements StripePaymentService {
+public class StripePaymentProdiverImpl implements StripePaymentProvider {
     private static final String YOUR_DOMAIN = "http://localhost:4242";
     private final PaymentService paymentService;
-    private final StripeConfig stripeConfig;
+    private final StripeProperties stripeProperties;
 
     public void createPaymentSession(Payment payment) {
-        Stripe.apiKey = stripeConfig.getSecretKey();
+        Stripe.apiKey = stripeProperties.getSecretKey();
         SessionCreateParams params =
                 SessionCreateParams.builder()
                         .setMode(SessionCreateParams.Mode.PAYMENT)
@@ -49,7 +47,7 @@ public class StripePaymentServiceImpl implements StripePaymentService {
     }
 
     private String createPrice(Payment payment) {
-        Stripe.apiKey = stripeConfig.getSecretKey();
+        Stripe.apiKey = stripeProperties.getSecretKey();
         Map<String, Object> params = new HashMap<>();
         params.put("unit_amount",
                 paymentService.calculatePrice(payment)
