+++ src/main/java/com/example/carsharingservice/stripepaymant/StripeProperties.java
@@ -1,4 +1,4 @@
-package com.example.carsharingservice.config;
+package com.example.carsharingservice.stripepaymant;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -8,12 +8,8 @@ import org.springframework.stereotype.Component;
 
 @Component
 @Getter
-@Setter
 @PropertySource("classpath:application.properties")
-public class StripeConfig {
-    @Value("${stripe.apikey}")
-    private String stripeKey;
-
+public class StripeProperties {
     @Value("${stripe.apikey}")
     private String secretKey;
 }
